修改数据库方言构建分页方法，参数类型

+++ buession-dao/src/main/java/com/buession/dao/mybatis/dialect/DB2Dialect.java
@@ -38,7 +38,7 @@ public final class DB2Dialect implements Dialect {
 	}
 
 	@Override
-	public String buildPaginationSql(final String sql, final long offset, final long limit) {
+	public String buildPaginationSql(final String sql, final int offset, final int limit) {
 		final StringBuilder sb = new StringBuilder("SELECT * FROM ");
 
 		sb.append("(SELECT TEMP.*, ROWNUMBER() OVER() AS ROW_ID FROM (").append(sql).append(") AS TEMP)");
